backend_Django/requirements.txt
Django>=4.2,<5
gunicorn>=23.0
psycopg2-binary>=2.9
python-dotenv>=1.0
